catuserbot version 2.9.3
CatUserbot Version 2.9.3
Complete info is done you can check by .info
 ☞ Updated python to 3.9.0
 ☞ Custom alive header and alive emoji's added
 ☞ CUSTOM_STICKER_PACKNAME  is added
 ☞ OWNER_ID is added to set the required account as owner
 ☞ Added pin support in pm too and also unpin option for both groups and pm
 ☞ For stoping autoprofile commands instead of restarting bot added end option by command
 ☞ dice_dart_ball is renamed to emojigames and slot machine is added
 ☞ Small changes in evaluators
 ☞ removed extdl and it install automatically after every restart
 ☞ small changes in scam
 ☞ ffmpeg is modified like that you can extract audio only and trim or video trim
 ☞ Added file summary option counts no of files either for user or in chat . thanks to @deleteduser420 and @SpEcHiDe
 ☞ fpost is fixed
 ☞ Pmpermit is modified and added variables like {mention} , {username} for CUSTOM_PMPERMIT_TEXT and also without mentioning user you can aprove user from any where by his username
 ☞ fixed webp format for remove background
 ☞ Translater is fixed
 ☞ Finally some minor bugs fixes
☞ Added Broadcast plugin .for more info check .info broadcast,.
☞ Pmpermit default text is changed and modified it.
☞ Added shutdown button to turn of heroku dyno.
☞To use aria module install the files from @catplugins or set the var PLUGIN_CHANNEL with this id 1001224206775
☞ Added  extra command to update and rename plugins each to send the file as document .
☞ Added .disapprove all or .da all to disapprove all approved pms
☞ Telegraph media and telegraph text can also used as .tgm and .tgt respectively.
 ☞ Font selection added in memify.

+++ userbot/plugins/forward.py
@@ -1,9 +1,25 @@
 import string
 
+from telethon.tl.types import Channel
+
 from ..utils import admin_cmd, edit_or_reply, sudo_cmd
 from . import CMD_HELP
 
+global msg_cache
 msg_cache = {}
+
+
+global groupsid
+groupsid = []
+
+
+async def all_groups_id(cat):
+    catgroups = []
+    async for dialog in cat.client.iter_dialogs():
+        entity = dialog.entity
+        if isinstance(entity, Channel) and entity.megagroup:
+            catgroups.append(entity.id)
+    return catgroups
 
 
 @bot.on(admin_cmd(pattern="frwd$"))
@@ -32,6 +48,8 @@ async def _(event):
 @bot.on(admin_cmd(pattern="resend$"))
 @bot.on(sudo_cmd(pattern="resend$", allow_sudo=True))
 async def _(event):
+    if event.fwd_from:
+        return
     try:
         await event.delete()
     except:
@@ -42,20 +60,32 @@ async def _(event):
     await event.respond(m)
 
 
-@bot.on(admin_cmd(pattern=r"fpost\s+(.*)"))
-@bot.on(sudo_cmd(pattern="fpost\s+(.*)", allow_sudo=True))
+@bot.on(admin_cmd(pattern=r"fpost (.*)"))
+@bot.on(sudo_cmd(pattern=r"fpost (.*)", allow_sudo=True))
 async def _(event):
+    if event.fwd_from:
+        return
+    global groupsid
+    global msg_cache
     await event.delete()
     text = event.pattern_match.group(1)
     destination = await event.get_input_chat()
+    if len(groupsid) == 0:
+        groupsid = await all_groups_id(event)
     for c in text.lower():
         if c not in string.ascii_lowercase:
             continue
         if c not in msg_cache:
-            async for msg in event.client.iter_messages(None, search=c):
+            async for msg in event.client.iter_messages(event.chat_id, search=c):
                 if msg.raw_text.lower() == c and msg.media is None:
                     msg_cache[c] = msg
                     break
+        if c not in msg_cache:
+            for i in groupsid:
+                async for msg in event.client.iter_messages(event.chat_id, search=c):
+                    if msg.raw_text.lower() == c and msg.media is None:
+                        msg_cache[c] = msg
+                        break
         await event.client.forward_messages(destination, msg_cache[c])
