+++ userbot/plugins/randomsticker.py
@@ -8,9 +8,8 @@ import requests
 from PIL import Image
 from telethon import functions, types, utils
 
-from userbot import CMD_HELP
-
 from ..utils import admin_cmd, sudo_cmd
+from . import CMD_HELP
 
 BASE_URL = "https://headp.at/pats/{}"
 PAT_IMAGE = "pat.webp"
@@ -19,6 +18,8 @@ PAT_IMAGE = "pat.webp"
 @bot.on(admin_cmd(pattern="cat$"))
 @bot.on(sudo_cmd(pattern="cat$", allow_sudo=True))
 async def _(event):
+    if event.fwd_from:
+        return
     try:
         await event.delete()
     except BaseException:
